Made 'Result' separate project
