+++ TagsCloudVisualization/ResultPatterLibrary/None.cs
@@ -1,0 +1,7 @@
+﻿namespace ResultPatterLibrary
+{
+    public class None
+    {
+        private None() { }
+    }
+}
