+++ TagsCloudVisualization/ResultPatterLibrary/Result.cs
@@ -1,15 +1,7 @@
 ﻿using System;
 
-
-namespace TagsCloudVisualization.Results
+namespace ResultPatterLibrary
 {
-    public class None
-    {
-        private None()
-        {
-        }
-    }
-
     public struct Result<T>
     {
         public Result(string error, T value = default)
@@ -82,8 +74,7 @@ namespace TagsCloudVisualization.Results
             return input;
         }
 
-        public static Result<TInput> ReplaceError<TInput>(this Result<TInput> input, 
-                                                          Func<string, string> replaceError)
+        public static Result<TInput> ReplaceError<TInput>(this Result<TInput> input, Func<string, string> replaceError)
             => input.IsSuccess ? input : Fail<TInput>(replaceError(input.Error));
 
         public static Result<TInput> RefineError<TInput>(this Result<TInput> input, string errorMessage)
